refactor: Move api_request and error mapper from ui to data/core
