+++ lib/data/core/api_request/api_request.dart
@@ -5,12 +5,12 @@ import 'package:flutter/foundation.dart';
 
 import 'package:http/http.dart' as http;
 import 'package:sport_matcher/config/api_config.dart';
-import 'package:sport_matcher/ui/core/utilities/api_request/api_exception.dart';
-import 'package:sport_matcher/ui/core/utilities/api_request/api_result.dart';
-import 'package:sport_matcher/ui/core/utilities/api_request/error_response.dart';
-import 'package:sport_matcher/ui/core/utilities/api_request/http_method.dart';
-import 'package:sport_matcher/ui/core/utilities/mapper/api_error_to_user_message_mapper.dart';
-import 'package:sport_matcher/ui/core/utilities/mapper/api_error_to_user_message_mapper_impl.dart';
+import 'package:sport_matcher/data/core/api_request/api_exception.dart';
+import 'package:sport_matcher/data/core/api_request/api_result.dart';
+import 'package:sport_matcher/data/core/api_request/error_response.dart';
+import 'package:sport_matcher/data/core/api_request/http_method.dart';
+import 'package:sport_matcher/data/core/mapper/abstract_api_error_to_user_message_mapper.dart';
+import 'package:sport_matcher/data/core/mapper/api_error_to_user_message_mapper.dart';
 
 class ApiRequest<TResponse> {
   final String path;
